feat : 친구 & 경쟁 가져오기 API 수정

+++ src/main/java/com/sunny/backend/competition/service/CompetitionService.java
@@ -3,15 +3,19 @@ package com.sunny.backend.competition.service;
 import static com.sunny.backend.competition.exception.CompetitionErrorCode.*;
 
 import java.time.Duration;
-import java.util.List;
 
 import javax.transaction.Transactional;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import com.sunny.backend.auth.jwt.CustomUserPrincipal;
 import com.sunny.backend.common.exception.CustomException;
+import com.sunny.backend.common.response.CommonResponse;
+import com.sunny.backend.common.response.ResponseService;
 import com.sunny.backend.competition.domain.Competition;
+import com.sunny.backend.competition.domain.CompetitionOutputStatus;
 import com.sunny.backend.competition.dto.request.CompetitionRequest;
 import com.sunny.backend.competition.dto.response.CompetitionStatusResponse;
 import com.sunny.backend.competition.repository.CompetitionRepository;
@@ -38,9 +42,10 @@ public class CompetitionService {
 	private final FriendNotiService friendNotiService;
 	private final UserRepository userRepository;
 	private final FriendCompetitionRepository friendCompetitionRepository;
+	private final ResponseService responseService;
 
 	@Transactional
-	public FriendCompetitionResponses applyCompetition(
+	public ResponseEntity<CommonResponse.SingleResponse<FriendCompetitionResponses>> applyCompetition(
 		CustomUserPrincipal customUserPrincipal,
 		CompetitionRequest competitionRequest
 	) {
@@ -49,7 +54,7 @@ public class CompetitionService {
 		friendWithUser.validateUser(users.getId());
 
 		// 이미 대결 신청을 보냈거나 받았거나 진행중인 경우 에러 발생
-		friendCompetitionRepository.findByFriendOrderByCreatedDateDesc(friendWithUser)
+		friendCompetitionRepository.findFirstByFriendOrderByCreatedDateDesc(friendWithUser)
 			.ifPresent(FriendCompetition::validateIsCompeting);
 
 		Friend friendWithUserFriend = friendRepository.findByUsersAndUserFriend(friendWithUser.getUserFriend(),
@@ -57,48 +62,63 @@ public class CompetitionService {
 			.orElseThrow(() -> new CustomException(FriendErrorCode.FRIEND_NOT_FOUND));
 
 		// 대결을 받은 사람이 신청한 사람과 경쟁중이 아닌데 경쟁중 상태로 뜨는 경우, 서버 이슈라 생각하고 해당 경쟁을 삭제
-		friendCompetitionRepository.findByFriendOrderByCreatedDateDesc(friendWithUserFriend)
+		friendCompetitionRepository.findFirstByFriendOrderByCreatedDateDesc(friendWithUserFriend)
 			.ifPresent(friendCompetition -> {
-				friendCompetitionRepository.delete(friendCompetition);
-				competitionRepository.delete(friendCompetition.getCompetition());
+				if (friendCompetition.isCompeting()) {
+					friendCompetitionRepository.delete(friendCompetition);
+					competitionRepository.delete(friendCompetition.getCompetition());
+				}
 			});
 
-		List<FriendCompetition> friendCompetitions = FriendCompetition.of(
-			friendWithUser,
-			friendWithUserFriend,
+		Competition competition = Competition.of(
 			competitionRequest.message(),
 			competitionRequest.startDate(),
 			competitionRequest.endDate(),
 			competitionRequest.price(),
 			competitionRequest.compensation()
 		);
-		friendCompetitionRepository.saveAll(friendCompetitions);
+		competitionRepository.save(competition);
+
+		FriendCompetition friendCompetition = FriendCompetition.of(friendWithUser, competition,
+			FriendCompetitionStatus.SEND,
+			CompetitionOutputStatus.NONE);
+		friendCompetitionRepository.save(friendCompetition);
+
+		FriendCompetition friendCompetitionToUserFriend = FriendCompetition.of(friendWithUserFriend, competition,
+			FriendCompetitionStatus.RECEIVE,
+			CompetitionOutputStatus.NONE);
+		friendCompetitionRepository.save(friendCompetitionToUserFriend);
 
 		String title = "[SUNNY] " + friendWithUser.getUsers().getNickname();
 		String body = "님으로부터 대결 신청을 받았어요.";
 		String bodyTitle = "대결 신청을 받았어요!";
 		friendNotiService.sendNotifications(title, body, bodyTitle, friendWithUser);
-		return FriendCompetitionResponses.from(friendCompetitions.get(0));
+		return responseService.getSingleResponse(HttpStatus.OK.value(),
+			FriendCompetitionResponses.from(friendCompetition), "");
 	}
 
-	@Transactional
-	public void acceptCompetition(CustomUserPrincipal customUserPrincipal, Long friendId) {
-		Users users = userRepository.getById(customUserPrincipal.getId());
+	public Friend updateCompetitionReturnFriend(
+		Long userId,
+		Long friendId,
+		FriendCompetitionStatus userStatus,
+		FriendCompetitionStatus userFriendStatus
+	) {
+		Users users = userRepository.getById(userId);
 		Friend friendWithUser = friendRepository.getById(friendId);
 		friendWithUser.validateUser(users.getId());
 
-		FriendCompetition friendCompetition = friendCompetitionRepository.findByFriendOrderByCreatedDateDesc(
+		FriendCompetition friendCompetition = friendCompetitionRepository.findFirstByFriendOrderByCreatedDateDesc(
 				friendWithUser)
 			.orElseThrow(() -> new CustomException(COMPETITION_NOT_FOUND));
 		if (!friendCompetition.isFriendCompetitionStatus(FriendCompetitionStatus.RECEIVE)) {
 			throw new CustomException(COMPETITION_NOT_RECEIVE);
 		}
-		friendCompetition.updateFriendCompetitionStatus(FriendCompetitionStatus.PROCEEDING);
+		friendCompetition.updateFriendCompetitionStatus(userStatus);
 
 		Friend friendWithUserFriend = friendRepository.findByUsersAndUserFriend(friendWithUser.getUserFriend(),
 				friendWithUser.getUsers())
 			.orElseThrow(() -> new CustomException(FriendErrorCode.FRIEND_NOT_FOUND));
-		FriendCompetition friendCompetitionUserFriend = friendCompetitionRepository.findByFriendOrderByCreatedDateDesc(
+		FriendCompetition friendCompetitionUserFriend = friendCompetitionRepository.findFirstByFriendOrderByCreatedDateDesc(
 				friendWithUserFriend)
 			.orElseThrow(() -> new CustomException(COMPETITION_NOT_FOUND));
 		// 대결을 받은 사람은 있는데 신청자가 없는 경우, 해당 대결을 삭제
@@ -108,7 +128,15 @@ public class CompetitionService {
 			competitionRepository.delete(friendCompetition.getCompetition());
 			throw new CustomException(COMPETITION_SERVER_ERROR);
 		}
-		friendCompetitionUserFriend.updateFriendCompetitionStatus(FriendCompetitionStatus.PROCEEDING);
+		friendCompetitionUserFriend.updateFriendCompetitionStatus(userFriendStatus);
+
+		return friendWithUser;
+	}
+
+	@Transactional
+	public void acceptCompetition(CustomUserPrincipal customUserPrincipal, Long friendId) {
+		Friend friendWithUser = updateCompetitionReturnFriend(customUserPrincipal.getId(), friendId,
+			FriendCompetitionStatus.PROCEEDING, FriendCompetitionStatus.PROCEEDING);
 
 		//TODO title,body,bodyTitle 분리
 		String title = "[SUNNY] " + friendWithUser.getUsers().getNickname();
@@ -119,32 +147,8 @@ public class CompetitionService {
 
 	@Transactional
 	public void refuseFriend(CustomUserPrincipal customUserPrincipal, Long friendId) {
-		Users users = userRepository.getById(customUserPrincipal.getId());
-		Friend friendWithUser = friendRepository.getById(friendId);
-		friendWithUser.validateUser(users.getId());
-
-		FriendCompetition friendCompetition = friendCompetitionRepository.findByFriendOrderByCreatedDateDesc(
-				friendWithUser)
-			.orElseThrow(() -> new CustomException(COMPETITION_NOT_FOUND));
-		if (!friendCompetition.isFriendCompetitionStatus(FriendCompetitionStatus.RECEIVE)) {
-			throw new CustomException(COMPETITION_NOT_RECEIVE);
-		}
-		friendCompetition.updateFriendCompetitionStatus(FriendCompetitionStatus.REFUSE);
-
-		Friend friendWithUserFriend = friendRepository.findByUsersAndUserFriend(friendWithUser.getUserFriend(),
-				friendWithUser.getUsers())
-			.orElseThrow(() -> new CustomException(FriendErrorCode.FRIEND_NOT_FOUND));
-		FriendCompetition friendCompetitionUserFriend = friendCompetitionRepository.findByFriendOrderByCreatedDateDesc(
-				friendWithUserFriend)
-			.orElseThrow(() -> new CustomException(COMPETITION_NOT_FOUND));
-		// 대결을 받은 사람은 있는데 신청자가 없는 경우, 해당 대결을 삭제
-		if (!friendCompetitionUserFriend.isFriendCompetitionStatus(FriendCompetitionStatus.SEND)) {
-			friendCompetitionRepository.delete(friendCompetition);
-			friendCompetitionRepository.delete(friendCompetitionUserFriend);
-			competitionRepository.delete(friendCompetition.getCompetition());
-			throw new CustomException(COMPETITION_SERVER_ERROR);
-		}
-		friendCompetitionUserFriend.updateFriendCompetitionStatus(FriendCompetitionStatus.REFUSE);
+		Friend friendWithUser = updateCompetitionReturnFriend(customUserPrincipal.getId(), friendId,
+			FriendCompetitionStatus.REFUSE, FriendCompetitionStatus.NONE);
 
 		String title = "[SUNNY] " + friendWithUser.getUsers().getNickname();
 		String body = "님이 대결을 거절했어요";
@@ -159,16 +163,24 @@ public class CompetitionService {
 		Friend friendWithUser = friendRepository.getById(friendId);
 		friendWithUser.validateUser(users.getId());
 
-		FriendCompetition friendCompetition = friendCompetitionRepository.findByFriendOrderByCreatedDateDesc(
+		FriendCompetition friendCompetition = friendCompetitionRepository.findFirstByFriendOrderByCreatedDateDesc(
 				friendWithUser)
 			.orElseThrow(() -> new CustomException(COMPETITION_NOT_FOUND));
 		if (!friendCompetition.isCompeting()) {
 			throw new CustomException(COMPETITION_NOT_FOUND);
 		}
 		friendCompetition.updateFriendCompetitionStatus(FriendCompetitionStatus.GIVE_UP);
+
+		Friend friendWithUserFriend = friendRepository.findByUsersAndUserFriend(friendWithUser.getUserFriend(),
+				friendWithUser.getUsers())
+			.orElseThrow(() -> new CustomException(FriendErrorCode.FRIEND_NOT_FOUND));
+		FriendCompetition friendCompetitionUserFriend = friendCompetitionRepository.findFirstByFriendOrderByCreatedDateDesc(
+				friendWithUserFriend)
+			.orElseThrow(() -> new CustomException(COMPETITION_NOT_FOUND));
+		friendCompetitionUserFriend.updateFriendCompetitionStatus(FriendCompetitionStatus.NONE);
 	}
 
-	public List<FriendCompetitionResponses> getCompetition(
+	public ResponseEntity<CommonResponse.ListResponse<FriendCompetitionResponses>> getCompetition(
 		CustomUserPrincipal customUserPrincipal,
 		Long friendId,
 		Long competitionId
@@ -177,14 +189,16 @@ public class CompetitionService {
 		Friend friend = friendRepository.getById(friendId);
 		friend.validateUser(users.getId());
 
-		return friendCompetitionRepository.getByFriendAndCompetition(friendId, competitionId)
-			.stream()
-			.map(FriendCompetitionResponses::from)
-			.toList();
+		return responseService.getListResponse(HttpStatus.OK.value(),
+			friendCompetitionRepository.getByFriendAndCompetition(friendId, competitionId)
+				.stream()
+				.map(FriendCompetitionResponses::from)
+				.toList(),
+			"");
 	}
 
 	@Transactional
-	public CompetitionStatusResponse getCompetitionStatus(
+	public ResponseEntity<CommonResponse.SingleResponse<CompetitionStatusResponse>> getCompetitionStatus(
 		CustomUserPrincipal customUserPrincipal,
 		Long friendId,
 		Long competitionId
@@ -205,10 +219,9 @@ public class CompetitionService {
 		double percentageUsed = calculateUserPercentage(user.getId(), competition);
 		double friendsPercentageUsed = calculateUserPercentage(userFriend.getId(), competition);
 
-		//	TODO 조회할 때 결과를 반영하는 것이 아니라, 소비할 때만 해도 될 것 같음
-		// competition.getOutput().updateOutput(percentageUsed, friendsPercentageUsed, user.getId(), userFriend.getId());
+		competition.getOutput().updateOutput(percentageUsed, friendsPercentageUsed, user.getId(), userFriend.getId());
 
-		return CompetitionStatusResponse.builder()
+		return responseService.getSingleResponse(HttpStatus.OK.value(), CompetitionStatusResponse.builder()
 			.competitionId(competition.getId())
 			.price(competition.getPrice())
 			.compensation(competition.getCompensation())
@@ -218,7 +231,7 @@ public class CompetitionService {
 			.friendName(userFriend.getNickname())
 			.userPercent(percentageUsed)
 			.friendsPercent(friendsPercentageUsed)
-			.build();
+			.build(), "");
 	}
 
 	//TODO 메소드 분리
